Add initial README file
This is an example of a commit message. A "good" commit message typically starts
with a short description on the first line, which will be used as the commit
'short log'. Then there must be an empty line, and finally a better description
which can be as long as needed.

Finally, the commit message should end with the author's signed off. While this
is not stricly required since the author identity is already recorder in the
commit object, most open source projects expects the signed-off-by tag to be
added.

Signed-off-by: Nicolas Dechesne <ndec@linux.com>